God, this codebase is getting COMPLICATED
